removed dependency on python humanize module.
The machine_audit callback plugin was importing the 'humanize' python module (not part of standard library).  Since this callback is invoked on all playbook runs, it's overkill to require anyone using a playbook to install this python module.  A simple refactor to add the relevant method directly to the plugin file.