Disable compaction_preheat_key_cache in cassandra
compaction_preheat_key_cache doesn't seem to fit our bulk loading or bulk indexing workflows. 

In the bulk loading scenario, once a record with a given key has been written to once, it isn't expected to be read again within the next few seconds, so if compaction occurs during that time it doesn't need to result in the key cache staying up to date.

In the bulk indexing scenario, if compaction occurs for any reason on some part of the database, it is likely a section that has already been read, so if it is compacted by the automatic compaction process, it isn't likely to be needed in the next few seconds again.