issue #187 : Add solr rebalanceleaders operation to monit after restarts
Need to rebalance leaders for shards in the solr cloud after every restart to ensure optimum load spread across nodes.

This adds that capability using monit.

We cannot use the alias directly for rebalanceleaders so we need to call listaliases, and then parse its json output with the jq program to identify the current biocache collection to rebalance

Signed-off-by: Peter Ansell <p_ansell@yahoo.com>